# use buffers for pathJoin, to re-use buffers.