--- a/cmd/object-api-utils_test.go
+++ b/cmd/object-api-utils_test.go
@@ -24,6 +24,7 @@ import (
 	"io"
 	"net/http"
 	"net/http/httptest"
+	"path"
 	"reflect"
 	"runtime"
 	"strconv"
@@ -35,6 +36,38 @@ import (
 	"github.com/minio/minio/internal/crypto"
 	"github.com/minio/pkg/trie"
 )
+
+func pathJoinOld(elem ...string) string {
+	trailingSlash := ""
+	if len(elem) > 0 {
+		if hasSuffixByte(elem[len(elem)-1], SlashSeparatorChar) {
+			trailingSlash = SlashSeparator
+		}
+	}
+	return path.Join(elem...) + trailingSlash
+}
+
+func BenchmarkPathJoinOld(b *testing.B) {
+	b.Run("PathJoin", func(b *testing.B) {
+		b.ResetTimer()
+		b.ReportAllocs()
+
+		for i := 0; i < b.N; i++ {
+			pathJoinOld("volume", "path/path/path")
+		}
+	})
+}
+
+func BenchmarkPathJoin(b *testing.B) {
+	b.Run("PathJoin", func(b *testing.B) {
+		b.ResetTimer()
+		b.ReportAllocs()
+
+		for i := 0; i < b.N; i++ {
+			pathJoin("volume", "path/path/path")
+		}
+	})
+}
 
 // Wrapper
 func TestPathTraversalExploit(t *testing.T) {